move mapMessageSenders to utils.matrix.ts
