+++ src/store/messages/saga.ts
@@ -25,9 +25,8 @@ import { Events as ChatEvents, getChatBus } from '../chat/bus';
 import { Uploadable, createUploadableFile } from './uploadable';
 import { chat } from '../../lib/chat';
 import { activeChannelIdSelector } from '../chat/selectors';
-import { featureFlags } from '../../lib/feature-flags';
 import { User } from '../channels';
-import { getZEROUsers as getZEROUsersAPI } from '../channels-list/api';
+import { mapMessageSenders } from './utils.matrix';
 
 export interface Payload {
   channelId: string;
@@ -93,7 +92,7 @@ const _isActive = (channelId) => (state) => {
 
 const FETCH_CHAT_CHANNEL_INTERVAL = 60000;
 
-function* getZeroUsersMap() {
+export function* getZeroUsersMap() {
   const users = yield select((state) => state.normalized.users);
   const zeroUsersMap: { [matrixId: string]: User } = {};
   for (const user of Object.values(users)) {
@@ -110,40 +109,6 @@ function* getZeroUsersMap() {
   } as User;
 
   return zeroUsersMap;
-}
-
-export function* mapMessageSenders(messages) {
-  if (!featureFlags.enableMatrix) {
-    return;
-  }
-
-  const zeroUsersMap = yield call(getZeroUsersMap);
-
-  const matrixIds = [];
-  for (const m of messages) {
-    // it's possible that our "cache" doesn't have the sender because (s)he has left the room
-    // in that case we need to record the Id, and fetch it's profile using the API
-    if (m.sender?.userId && !zeroUsersMap[m.sender.userId]) {
-      matrixIds.push(m.sender.userId);
-    }
-  }
-
-  if (!matrixIds.length) {
-    const zeroUsers = yield call(getZEROUsersAPI, matrixIds);
-    for (const user of zeroUsers) {
-      zeroUsersMap[user.matrixId] = {
-        userId: user.id,
-        profileId: user.profileSummary?.id,
-        firstName: user.profileSummary?.firstName,
-        lastName: user.profileSummary?.lastName,
-        profileImage: user.profileSummary?.profileImage,
-      };
-    }
-  }
-
-  messages.forEach((message) => {
-    message.sender = zeroUsersMap[message.sender?.userId] || message.sender;
-  });
 }
 
 export function* fetch(action) {
